refactor: move cs controller and pass store in constructor
