+++ src/content-scripts/index.js
@@ -1,14 +1,16 @@
 import store from '@/store'
 
 import Overlay from '@/modules/overlay'
-import Recorder from '@/modules/recorder'
 import Shooter from '@/modules/shooter'
-import Controller from '@/modules/controller'
+import Recorder from '@/modules/recorder'
+
+import Controller from '@/content-scripts/controller'
 
 window.headlessRecorder = new Controller({
   shooter: new Shooter(),
   overlay: new Overlay({ store }),
   recorder: new Recorder({ store }),
+  store,
 })
 
 window.headlessRecorder.init()
